requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
jinja2
google-generativeai
